# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCConnect API",
    "version": "2.0.0",
    "description": "Authoritative REST contract for PCConnect controllers and agents. SignalR notifications are advisory; REST resources are authoritative."
  },
  "servers": [
    { "url": "https://api.pcconnect.adamdeveloping.co.uk/api/v2", "description": "Production" }
  ],
...
  "paths": {
...
    "/auth/password/login": {
      "post": {
        "tags": ["Identity"], "operationId": "passwordLogin", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordLoginRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/TokenPair" }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Identity"], "operationId": "registerAccount", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegistrationRequest" } } } },
        "responses": { "202": { "description": "Account created; email verification required" }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Identity"], "operationId": "refreshSession", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/TokenPair" }, "401": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Identity"], "operationId": "logout", "responses": { "204": { "description": "Session revoked" }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/auth/passkeys/registration/options": {
      "post": {
        "tags": ["Identity"], "operationId": "createPasskeyRegistrationOptions",
        "responses": { "200": { "description": "WebAuthn creation options", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebAuthnOptions" } } } }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/passkeys/registration/complete": {
      "post": {
        "tags": ["Identity"], "operationId": "completePasskeyRegistration", "parameters": [{ "$ref": "#/components/parameters/StepUpGrant" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebAuthnCredential" } } } },
        "responses": { "201": { "description": "Passkey registered", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Passkey" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/passkeys/{passkeyId}": {
      "delete": {
        "tags": ["Account"], "operationId": "removePasskey", "parameters": [{ "$ref": "#/components/parameters/PasskeyId" }, { "$ref": "#/components/parameters/StepUpGrant" }],
        "responses": { "204": { "description": "Passkey revoked" }, "403": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/passkeys/authentication/options": {
      "post": {
        "tags": ["Identity"], "operationId": "createPasskeyAuthenticationOptions", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasskeyAuthenticationOptionsRequest" } } } },
        "responses": { "200": { "description": "WebAuthn request options", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebAuthnOptions" } } } }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/passkeys/authentication/complete": {
      "post": {
        "tags": ["Identity"], "operationId": "completePasskeyAuthentication", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WebAuthnCredential" } } } },
        "responses": { "200": { "$ref": "#/components/responses/TokenPair" }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/device-enrollments": {
      "post": {
        "tags": ["Enrollment"], "operationId": "createDeviceEnrollment", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceEnrollmentRequest" } } } },
        "responses": { "201": { "description": "Pending enrollment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceEnrollment" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/device-enrollments/{userCode}/approve": {
      "post": {
        "tags": ["Enrollment"], "operationId": "approveDeviceEnrollment", "parameters": [{ "$ref": "#/components/parameters/UserCode" }],
        "responses": { "204": { "description": "Enrollment approved" }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/device-enrollments/token": {
      "post": {
        "tags": ["Enrollment"], "operationId": "exchangeDeviceCode", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceCodeRequest" } } } },
        "responses": { "200": { "description": "Approved device credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceTokenPair" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/agent/auth/refresh": {
      "post": {
        "tags": ["Agent"], "operationId": "refreshDeviceCredential", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefreshRequest" } } } },
        "responses": { "200": { "description": "Rotated device credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceTokenPair" } } } }, "401": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/devices": {
      "get": {
        "tags": ["Devices"], "operationId": "listDevices", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "description": "Owned devices", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DevicePage" } } } } }
      }
    },
    "/devices/{deviceId}": {
      "get": { "tags": ["Devices"], "operationId": "getDevice", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }], "responses": { "200": { "description": "Device", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Device" } } } }, "404": { "$ref": "#/components/responses/Problem" } } },
      "patch": {
        "tags": ["Devices"], "operationId": "updateDevice", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
        "requestBody": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/DeviceUpdate" } } } },
        "responses": { "200": { "description": "Updated device", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Device" } } } }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      },
      "delete": { "tags": ["Devices"], "operationId": "revokeDevice", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }, { "$ref": "#/components/parameters/StepUpGrant" }], "responses": { "204": { "description": "Device and credentials revoked" }, "403": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" } } }
    },
    "/devices/{deviceId}/windows-sids": {
      "get": {
        "tags": ["Devices"], "operationId": "listDeviceWindowsSids", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
        "responses": { "200": { "description": "Authorized and pending interactive Windows identities", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WindowsSidStatus" } } } } }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/devices/{deviceId}/windows-sids/{windowsSid}/authorize": {
      "post": {
        "tags": ["Devices"], "operationId": "authorizeDeviceWindowsSid", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }, { "$ref": "#/components/parameters/WindowsSid" }, { "$ref": "#/components/parameters/StepUpGrant" }],
        "responses": { "204": { "description": "Windows identity authorized" }, "403": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/devices/{deviceId}/windows-sids/{windowsSid}": {
      "delete": {
        "tags": ["Devices"], "operationId": "revokeDeviceWindowsSid", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }, { "$ref": "#/components/parameters/WindowsSid" }, { "$ref": "#/components/parameters/StepUpGrant" }],
        "responses": { "204": { "description": "Windows identity authorization revoked" }, "403": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/agent/heartbeat": {
      "post": {
        "tags": ["Agent"], "operationId": "heartbeat", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Heartbeat" } } } },
        "responses": { "204": { "description": "Presence updated" }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/agent/windows-sid-candidates": {
      "post": {
        "tags": ["Agent"], "operationId": "registerWindowsSidCandidate",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WindowsSidCandidateRequest" } } } },
        "responses": { "200": { "description": "Current candidate or authorization state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WindowsSidStatus" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque" }
    },
    "parameters": {
      "Cursor": { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "maxLength": 512 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
      "DeviceId": { "name": "deviceId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
      "WindowsSid": { "name": "windowsSid", "in": "path", "required": true, "schema": { "type": "string", "minLength": 5, "maxLength": 184, "pattern": "^S-1-(?:[0-9]+-){1,14}[0-9]+$" } },
...
      "PasskeyId": { "name": "passkeyId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "UserCode": { "name": "userCode", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^[A-HJ-NP-Z2-9]{8}$" } },
...
      "StepUpGrant": { "name": "X-Step-Up-Grant", "in": "header", "required": true, "schema": { "type": "string", "minLength": 32, "maxLength": 512 } },
...
    },
    "responses": {
      "Problem": { "description": "RFC 9457 problem details", "headers": { "X-Correlation-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "TokenPair": { "description": "Rotated user session credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenPair" } } } }
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Instant": { "type": "string", "format": "date-time" },
...
      "Problem": { "type": "object", "additionalProperties": true, "required": ["type", "title", "status", "code", "correlationId"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "detail": { "type": "string" }, "instance": { "type": "string", "format": "uri-reference" }, "code": { "type": "string", "pattern": "^[a-z0-9_]+$" }, "correlationId": { "type": "string" }, "errors": { "type": "array", "items": { "type": "object", "required": ["pointer", "code"], "properties": { "pointer": { "type": "string" }, "code": { "type": "string" } } } } } },
      "PasswordLoginRequest": { "type": "object", "additionalProperties": false, "required": ["login", "password", "client"], "properties": { "login": { "type": "string", "minLength": 1, "maxLength": 320 }, "password": { "type": "string", "minLength": 1, "maxLength": 1024, "format": "password" }, "client": { "$ref": "#/components/schemas/ClientDescriptor" } } },
      "RegistrationRequest": { "type": "object", "additionalProperties": false, "required": ["username", "email", "displayName", "password", "timezone", "marketingOptIn", "client"], "properties": { "username": { "type": "string", "minLength": 3, "maxLength": 100 }, "email": { "type": "string", "format": "email", "maxLength": 320 }, "displayName": { "type": "string", "minLength": 1, "maxLength": 100 }, "password": { "type": "string", "minLength": 12, "maxLength": 1024, "format": "password" }, "timezone": { "type": "string", "minLength": 1, "maxLength": 100 }, "dateOfBirth": { "type": ["string", "null"], "format": "date" }, "marketingOptIn": { "type": "boolean" }, "client": { "$ref": "#/components/schemas/ClientDescriptor" } } },
      "ClientDescriptor": { "type": "object", "additionalProperties": false, "required": ["platform", "name", "version"], "properties": { "platform": { "$ref": "#/components/schemas/Platform" }, "name": { "type": "string", "minLength": 1, "maxLength": 100 }, "version": { "type": "string", "minLength": 1, "maxLength": 40 } } },
      "RefreshRequest": { "type": "object", "additionalProperties": false, "required": ["refreshToken"], "properties": { "refreshToken": { "type": "string", "minLength": 43, "maxLength": 512, "format": "password" } } },
      "TokenPair": { "type": "object", "additionalProperties": false, "required": ["accessToken", "accessTokenExpiresAt", "refreshToken", "refreshTokenExpiresAt", "sessionId"], "properties": { "accessToken": { "type": "string", "format": "password" }, "accessTokenExpiresAt": { "$ref": "#/components/schemas/Instant" }, "refreshToken": { "type": "string", "format": "password" }, "refreshTokenExpiresAt": { "$ref": "#/components/schemas/Instant" }, "sessionId": { "$ref": "#/components/schemas/Uuid" } } },
...
      "WebAuthnOptions": { "type": "object", "description": "JSON-serialized PublicKeyCredential creation or request options plus a server challenge ID.", "additionalProperties": true, "required": ["challengeId", "publicKey"], "properties": { "challengeId": { "$ref": "#/components/schemas/Uuid" }, "publicKey": { "type": "object" } } },
      "WebAuthnCredential": { "type": "object", "additionalProperties": false, "required": ["challengeId", "credential"], "properties": { "challengeId": { "$ref": "#/components/schemas/Uuid" }, "credential": { "type": "object", "additionalProperties": true } } },
      "PasskeyAuthenticationOptionsRequest": { "type": "object", "additionalProperties": false, "required": ["client"], "properties": { "loginHint": { "type": "string", "maxLength": 320 }, "client": { "$ref": "#/components/schemas/ClientDescriptor" } } },
      "Passkey": { "type": "object", "additionalProperties": false, "required": ["id", "displayName", "createdAt"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "displayName": { "type": "string", "minLength": 1, "maxLength": 100 }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "lastUsedAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] } } },
...
      "Platform": { "type": "string", "enum": ["windows", "android", "ios", "macos", "linux", "web", "unknown"] },
      "DeviceCapability": { "type": "string", "enum": ["lock", "sleep", "hibernate", "sign_out", "restart", "shutdown", "reminders"] },
      "DeviceEnrollmentRequest": { "type": "object", "additionalProperties": false, "required": ["platform", "displayName", "agentVersion", "protocolVersion", "capabilities"], "properties": { "platform": { "$ref": "#/components/schemas/Platform" }, "displayName": { "type": "string", "minLength": 1, "maxLength": 100 }, "agentVersion": { "type": "string", "maxLength": 40 }, "protocolVersion": { "type": "integer", "minimum": 2 }, "timezone": { "type": "string", "maxLength": 100 }, "capabilities": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/DeviceCapability" } } } },
      "DeviceEnrollment": { "type": "object", "additionalProperties": false, "required": ["deviceCode", "userCode", "verificationUri", "expiresAt", "pollIntervalSeconds"], "properties": { "deviceCode": { "type": "string", "format": "password" }, "userCode": { "type": "string" }, "verificationUri": { "type": "string", "format": "uri" }, "expiresAt": { "$ref": "#/components/schemas/Instant" }, "pollIntervalSeconds": { "type": "integer", "minimum": 5 } } },
      "DeviceCodeRequest": { "type": "object", "additionalProperties": false, "required": ["deviceCode"], "properties": { "deviceCode": { "type": "string", "format": "password" } } },
      "DeviceTokenPair": { "type": "object", "additionalProperties": false, "required": ["deviceId", "accessToken", "accessTokenExpiresAt", "refreshToken", "refreshTokenExpiresAt"], "properties": { "deviceId": { "$ref": "#/components/schemas/Uuid" }, "accessToken": { "type": "string", "format": "password" }, "accessTokenExpiresAt": { "$ref": "#/components/schemas/Instant" }, "refreshToken": { "type": "string", "format": "password" }, "refreshTokenExpiresAt": { "$ref": "#/components/schemas/Instant" } } },
      "Device": { "type": "object", "additionalProperties": false, "required": ["id", "platform", "displayName", "agentVersion", "protocolVersion", "capabilities", "status", "lastSeenAt", "createdAt", "version"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "platform": { "$ref": "#/components/schemas/Platform" }, "displayName": { "type": "string" }, "agentVersion": { "type": "string" }, "protocolVersion": { "type": "integer" }, "timezone": { "type": ["string", "null"] }, "capabilities": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/DeviceCapability" } }, "status": { "type": "string", "enum": ["online", "offline", "revoked"] }, "lastSeenAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "version": { "type": "integer", "minimum": 1 } } },
      "DeviceUpdate": { "type": "object", "additionalProperties": false, "minProperties": 1, "properties": { "displayName": { "type": "string", "minLength": 1, "maxLength": 100 }, "expectedVersion": { "type": "integer", "minimum": 1 } } },
      "WindowsSidCandidateRequest": { "type": "object", "additionalProperties": false, "required": ["windowsSid"], "properties": { "windowsSid": { "type": "string", "minLength": 5, "maxLength": 184, "pattern": "^S-1-(?:[0-9]+-){1,14}[0-9]+$" }, "displayLabel": { "type": ["string", "null"], "maxLength": 100 } } },
      "WindowsSidStatus": { "type": "object", "additionalProperties": false, "required": ["windowsSid", "status"], "properties": { "windowsSid": { "type": "string", "minLength": 5, "maxLength": 184 }, "displayLabel": { "type": ["string", "null"] }, "status": { "type": "string", "enum": ["pending", "authorized"] }, "observedAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "authorizedAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] } } },
      "DevicePage": { "allOf": [{ "$ref": "#/components/schemas/Page" }, { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Device" } } } }] },
...
      "Heartbeat": { "type": "object", "additionalProperties": false, "required": ["agentInstanceId", "agentVersion", "protocolVersion", "capabilities", "sentAt"], "properties": { "agentInstanceId": { "$ref": "#/components/schemas/Uuid" }, "agentVersion": { "type": "string" }, "protocolVersion": { "type": "integer", "minimum": 2 }, "capabilities": { "type": "array", "uniqueItems": true, "items": { "$ref": "#/components/schemas/DeviceCapability" } }, "sentAt": { "$ref": "#/components/schemas/Instant" } } },
...
      "Page": { "type": "object", "additionalProperties": false, "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": true }, "nextCursor": { "type": ["string", "null"] } } }
    }
  },
  "security": [{ "bearerAuth": [] }]
}